Integration of new ViewAssembly. First cut.
Still have some work to do, so we'll see.
