+++ choam/src/test/resources/logback-test.xml
@@ -28,9 +28,6 @@
     <logger name="com.salesforce.apollo.choam.ViewAssembly" level="info" additivity="false">
         <appender-ref ref="STDOUT" />
     </logger>
-    <logger name="com.salesforce.apollo.choam.ViewAssembly2" level="info" additivity="false">
-        <appender-ref ref="STDOUT" />
-    </logger>
     <logger name="com.salesforce.apollo.choam.Producer" level="info" additivity="false">
         <appender-ref ref="STDOUT" />
     </logger>
